+++ revit-mcp-commandset/Services/TagWallsEventHandler.cs
@@ -224,7 +224,7 @@ try
             {
                 if (int.TryParse(_tagTypeId, out int id))
                 {
-                    ElementId elementId = new ElementId(id);
+                    ElementId elementId = new ElementId((long)id);
                     Element element = doc.GetElement(elementId);
 
                     if (element != null && element is FamilySymbol symbol &&
